+++ src/app/test-sample.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { add } from './sample'
+import { add } from './test-sample'
 
 describe('add', () => {
   it('1 + 2 = 3', () => {
